feat: 온보딩 화면 버튼 클릭 이벤트 구현
