+++ app/src/main/java/com/project/meongcare/onboarding/view/CompleteOnBoardingFragment.kt
@@ -1,0 +1,20 @@
+package com.project.meongcare.onboarding.view
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.project.meongcare.R
+
+class CompleteOnBoardingFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_complete_on_boarding, container, false)
+    }
+
+}
