+++ app/src/main/java/com/project/meongcare/onboarding/view/DogAddOnBoardingFragment.kt
@@ -1,0 +1,22 @@
+package com.project.meongcare.onboarding.view
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.project.meongcare.R
+
+class DogAddOnBoardingFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_dog_add_on_boarding, container, false)
+    }
+
+
+}
