+++ app/src/main/java/com/project/meongcare/onboarding/view/SecondOnBoardingFragment.kt
@@ -1,17 +1,18 @@
-package com.project.meongcare
+package com.project.meongcare.onboarding.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.project.meongcare.R
 
-class FirstOnBoardingFragment : Fragment() {
+class SecondOnBoardingFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.fragment_first_on_boarding, container, false)
+        return inflater.inflate(R.layout.fragment_second_on_boarding, container, false)
     }
 }
