+++ app/src/main/java/com/project/meongcare/onboarding/view/ThirdOnBoardingFragment.kt
@@ -1,10 +1,11 @@
-package com.project.meongcare
+package com.project.meongcare.onboarding.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.project.meongcare.R
 
 class ThirdOnBoardingFragment : Fragment() {
     override fun onCreateView(
