+++ app/src/main/res/layout/fragment_complete_on_boarding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CompleteOnBoardingFragment" >
+    tools:context=".onboarding.view.CompleteOnBoardingFragment" >
 
     <ImageView
         android:id="@+id/imageView5"
