+++ app/src/main/res/layout/fragment_first_on_boarding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".FirstOnBoardingFragment">
+    tools:context=".onboarding.view.FirstOnBoardingFragment">
 
     <TextView
         android:id="@+id/textView"
