+++ app/src/main/res/layout/fragment_second_on_boarding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".SecondOnBoardingFragment">
+    tools:context=".onboarding.view.SecondOnBoardingFragment">
 
     <TextView
         android:id="@+id/textView"
