+++ app/src/main/res/layout/fragment_third_on_boarding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".ThirdOnBoardingFragment">
+    tools:context=".onboarding.view.ThirdOnBoardingFragment">
 
     <TextView
         android:id="@+id/textView"
